blog-posts/ai-product-template/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Simple FastAPI service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.122.0",
    "uvicorn==0.38.0",
    "sqlalchemy==2.0.44",
    "pydantic==2.12.5",
    "pytest==9.0.1",
    "httpx==0.28.1",
    "python-jose[cryptography]==3.5.0",
]
